+++ server/microservices/info/controller.js
@@ -1,0 +1,14 @@
+const model = require('./model')
+const constants = require('../../util/constants')
+
+module.exports = {
+
+     async getInfo(req, res) {
+
+        //Obtenemos la promesa
+        var result = await model.getInfo()
+        console.log(result)
+
+        return res.send(result)
+    }
+}