+++ server/microservices/info/routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const controller = require('./controller')
+
+const router = express.Router();
+
+router.get('/info', controller.getInfo)
+
+module.exports = router